// packages/agent-provider-pi/src/pi-sdk/types.ts
export interface CustomOverlayOptions {
  anchor?: "bottom-center";
  width?: string;
  maxHeight?: string;
}